pyproject.toml
[project]
name = "sqlagent"
version = "0.1.0"
description = ""
authors = [
    {name = "Anwar Hermuche",email = "anwarhermuche2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "duckdb (>=1.2.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langgraph (>=0.4.5,<0.5.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)"
]

[tool.poetry]
packages = [{include = "sqlagent", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
